third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/report-only-in-meta.html

<!DOCTYPE html>
<html>
<head>
    <script>
        if (window.testRunner)
            testRunner.dumpAsText();
    </script>
    <meta http-equiv="Content-Security-Policy-Report-Only" content="sandbox allow-scripts; frame-ancestors 'none'; report-uri http://foo.test">
</head>
<body>
</body>
</html>
