third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/report-only-in-meta-expected.txt

CONSOLE ERROR: line 8: The report-only Content Security Policy 'sandbox allow-scripts; frame-ancestors 'none'; report-uri http://foo.test' was delivered via a <meta> element, which is disallowed. The policy has been ignored.
